+++ sources/BaseDestination.swift
@@ -126,8 +126,8 @@ public class BaseDestination: Hashable, Equatable {
     // XCode RGB colors
     public struct LevelColor {
         public var Verbose = "fg200,200,200;"     // silver
-        public var Debug = "fg0,0,255;"           // blue
-        public var Info = "fg0,255,0;"            // green
+        public var Debug = "fg0,255,0;"           // green
+        public var Info = "fg0,0,255;"           // blue
         public var Warning = "fg255,255,0;"       // yellow
         public var Error = "fg255,0,0;"           // red
     }